requirements.txt
numpy~=1.24.2
matplotlib~=3.7.1
Pillow
progressbar2
readlif
scikit-image~=0.20.0
scikit-learn
seaborn
